Restrict maximum width of GoAT diagrams
This prevents the diagrams from expanding to the full width of the
screen, which can be quite a lot of space, especially for large
viewports such as 1080p screens (desktop) and higher.

As this affects only the maximum width, the diagrams will still scale
down gracefully for mobile and other small viewports.